move tabindex setup to component creation
Instead of changing the tabindex onSpotlightFocused, set any spottable
control with a tabindex of -1 (programmatically focusable) on create.
This required moving the hacky Component.create code from container
to spotlight so we only hacked it once for this lib.

Issue: ENYO-2430
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)
